refactor: keep offline hook outside, add NotFound story
