+++ agir/front/components/allPages/NotFoundPage.js
@@ -1,15 +1,15 @@
 import React, { useEffect } from "react";
 
-import illustration from "./illustration.svg";
 import styled from "styled-components";
-
-import { useIsOffline } from "./hooks";
-
-import Button from "@agir/front/genericComponents/Button";
-import background from "@agir/front/genericComponents/images/illustration-404.svg";
 import style from "@agir/front/genericComponents/_variables.scss";
 
+import background from "@agir/front/genericComponents/images/illustration-404.svg";
+import illustration from "@agir/front/allPages/offline/illustration.svg";
+import { useIsOffline } from "@agir/front/allPages/offline/hooks";
+
 import TopBar from "@agir/front/allPages/TopBar";
+import Button from "@agir/front/genericComponents/Button";
+import Spacer from "@agir/front/genericComponents/Spacer";
 
 const Container = styled.div`
   display: flex;
@@ -18,7 +18,6 @@ const Container = styled.div`
   justify-content: center;
   width: 100%;
   min-height: calc(100vh - 74px);
-  margin-top: 74px;
   position: relative;
   overflow: auto;
   background-image: url("${background}");
@@ -65,7 +64,7 @@ const OfflineBlock = styled.div`
 `;
 
 // Most of the time, if we are in the app on unknow URL, it is because of service worker no network handling
-const NotFoundPage = () => {
+export const NotFoundPage = ({ isTopBar = true }) => {
   const isOffline = useIsOffline();
 
   // We try to reload every 5 second
@@ -83,7 +82,12 @@ const NotFoundPage = () => {
 
   return (
     <PageStyle>
-      <TopBar />
+      {isTopBar && (
+        <>
+          <TopBar />
+          <Spacer size="74px" />
+        </>
+      )}
       <Container>
         {isOffline ? (
           <OfflineBlock>
