+++ agir/front/components/allPages/NotFoundPage.stories.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { NotFoundPage } from "./NotFoundPage";
+
+export default {
+  component: NotFoundPage,
+  title: "Layout/NotFoundPage",
+};
+
+const Template = (args) => <NotFoundPage {...args} />;
+
+export const Default = Template.bind({});
+Default.args = { isTopBar: false };
